refactor: centralize saved display settings
